Clean up WebPoxy file, start to file cache
Wholesale changes were made to the WebProxy file to streamline and clean
up the code. Helper methods were added, along with documentation. Work
still could be done to make the logical structure better.

A file cache class has now been created and this will eventually
determine if a request can be served from file, and if it can it will
respond with an HTTP response and the file data.