Implement all responses as binary
Now, all HTTP requests are still taken as strings to allow easy
processing (checking validity, etc.).

But when reading responses from the desired server, it is read in binary
form. This binary data is then sent back to the client unchanged (to all
objects such as PDFs to be viewed in browser), and also cached as a
binary file.

As the pages are cached in binary form, this allows for simple file
serving from cache that allows binary files such as PDFs to be read by
the browser.